Finally working ManagerTest but @Before don't print
